requirements.txt
cassandra-driver
pydantic
fastapi[all]==0.115.5
uvicorn[standard]==0.32.1
requests
cryptography
